KEEP TESTING SORT 4

+++ mini_sort.c
@@ -6,7 +6,7 @@
 /*   By: gacorrei <gacorrei@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/09 15:55:44 by gacorrei          #+#    #+#             */
-/*   Updated: 2023/01/09 16:33:12 by gacorrei         ###   ########.fr       */
+/*   Updated: 2023/01/09 16:37:47 by gacorrei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,12 +43,12 @@ void	sort_3(t_list **stack)
 
 int	count_moves(t_list *stack, long long min)
 {
-	int	moves;
+	long long	moves;
 
 	moves = 0;
 	while (stack)
 	{
-		if (stack->content == min)
+		if ((long long)stack->content == min)
 			break ;
 		moves++;
 		stack = stack->next;
@@ -60,7 +60,7 @@ void	sort_4(t_list **stack_a, t_list **stack_b)
 {
 	t_list		*temp;
 	long long	min;
-	int			moves;
+	long long	moves;
 
 	min = find_min(*stack_a);
 	temp = *stack_a;
